requirements.txt
streamlit>=1.36
langchain>=0.3.27
langchain-openai>=0.3.30
google-generativeai>=0.8.0
python-dotenv>=1.0.0